Fix : third profile page layout
