+++ src/component/ContentCard/ContentCard.module.scss
@@ -1,7 +1,7 @@
 .content-box {
-  width: 240px;
-  height: 300px;
-  background-color: #161616;
+  width: 500px;
+  min-height: 300px;
+  background-color: #741f90;
   border-radius: 8px;
   padding: 20px 24px;
   display: flex;
