+++ src/component/PaintBoard/PaintBoard.module.scss
@@ -1,0 +1,102 @@
+.canvas-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 4px;
+  width: 600px;
+  height: 400px;
+
+  .tool-bar {
+    width: 100%;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+
+    .action-tools,
+    .revert-tools {
+      display: flex;
+      background-color: #cacaca;
+      gap: 4px;
+      padding: 4px;
+      width: fit-content;
+      border-radius: 4px;
+
+      button {
+        border: none;
+        border-radius: 4px;
+        cursor: pointer;
+        background-color: #cacaca;
+
+        &.tool:hover {
+          background-color: #a5a5a5;
+        }
+
+        &:disabled {
+          background-color: #fff;
+          cursor: default;
+        }
+      }
+    }
+
+    .tool {
+      width: fit-content;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      padding: 4px;
+
+      .icon {
+        width: 16px;
+        height: 16px;
+      }
+    }
+
+    .download-button {
+      width: fit-content;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      padding: 0;
+
+      .icon {
+        width: 24px;
+        height: 24px;
+      }
+    }
+
+    .palette {
+      width: fit-content;
+      background-color: #ffffff;
+      display: flex;
+      align-items: center;
+      justify-content: center;
+      gap: 8px;
+      padding: 0 8px;
+      border-radius: 4px;
+      box-sizing: content-box;
+
+      .color {
+        width: 16px;
+        height: 16px;
+        border-radius: 50%;
+        display: flex;
+        align-items: center;
+        justify-content: center;
+        cursor: pointer;
+
+        &.selected {
+          border: 2px solid transparent;
+
+          &::after {
+            content: "";
+            width: 80%;
+            height: 80%;
+            border-radius: 50%;
+            border: 1px solid rgb(255, 255, 255);
+          }
+        }
+      }
+    }
+  }
+}
