+++ src/component/WideCard/WideCard.module.scss
@@ -1,0 +1,47 @@
+.container {
+  width: 320px;
+  background-color: rgb(52, 52, 52);
+  border-radius: 4px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: start;
+
+  gap: 20px;
+  padding: 12px;
+  height: fit-content;
+  margin: 20px;
+
+  .content {
+    display: flex;
+    flex-direction: column;
+    align-items: start;
+    justify-content: center;
+    gap: 4px;
+
+    .title {
+      font-size: 20px;
+      font-weight: bold;
+      text-align: center;
+    }
+
+    .description {
+      font-size: 12px;
+    }
+
+    a {
+      color: #fff;
+      font-size: 12px;
+      align-self: flex-end;
+      margin-top: 12px;
+
+      &:hover {
+        color: yellow;
+      }
+    }
+  }
+
+  .video {
+    width: 100%;
+  }
+}
